+++ bin/synapseed/src/main.rs
@@ -1,3 +1,4 @@
+#![forbid(unsafe_code)]
 use std::path::{Path, PathBuf};
 
 use anyhow::Result;